Update game_over function
The function now works with a disk object instead of a player
object. It takes in a disk and compares that disk with other
disk objects around it. If four of the disk objects are the same
and in a row, it will return true. This will indicate if the game
is won by a player.