Remove redundant class
'c4_view.py' was supposed to be renamed to 'c4view.py'; but as a mistake
without realizing it 'c4view.py' happened to be a new file created.
This means that 'c4_ciew.py' is useless since the first changes to
the view class are in 'c4view.py'.