Fix and optimize GUI
The GUI from a previous commit was a 7x6 size board when it was
supposed to be a 6x7 size board. This change has to be made as
soon as possible because the model was initialized on the basis
that the board was 6x7 sized. Without this change the model would
not be able to interact with the view properly.