Add functionality design to game
In order to test the code written so far i had to comment out any disk-related
code in other classes because there is an error in the disk class from
someone elses commit.

Added a layout on how the game should function and be played in the controller.
I am making this change because someone else who would be handed the task of
finishing off the functionality of the game can just work off of the
layout that has already been created. The side effect of this commit is that
the current "event" will be printed on the shell when the program is run.