Add Documentation and create_stage
Added documentation to the methods that I created in previous commits.
Also added the create_stage method which creates the game board and
returns it. This change must be made in order for the game board to
be functional and the documentation was added to insure other know
what the methods do.