First step to connect view and model
Added colour constants to player, commented out code for a later commit,
column amount to model, getColors to model, and update to model.
Changed getColor in disk.py to use player in order to get the rgb values of a color.
Changed C4's draw_stage for loop for disks. Changed turn() in model.
Changed getPlayer# in player.